+++ Omise.Net/Enums/DisputeStatus.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Omise
+{
+    public enum DisputeStatus
+    {
+        Open,
+        Pending,
+        Won,
+        Lost
+    }
+}