+++ Omise.Net/Factories/DisputeFactory.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Omise
+{
+    public class DisputeFactory : GenericFactory<Dispute>
+    {
+        /// <summary>
+        /// Initialize the factory
+        /// </summary>
+        public DisputeFactory() { 
+            
+        }
+    }
+}